Add gh-pages workflow (#5)
* Add gh-pages workflow

This workflow will deploy the blog on push to the main branch.

It will build hugo from the root and then publish it to ./public on selected branch (currently default value gh-pages)

By default, the GitHub action pushes the generated content to the gh-pages branch. This means GitHub has to serve your gh-pages branch as a GitHub Pages branch. You can change this setting by going to Settings > GitHub Pages, and change the source branch to gh-pages